+++ hephaistos-data/lua/FixLuaState/QuestData.lua
@@ -1,2 +1,2 @@
 ScreenData.QuestLog.ItemStartX = Hephaistos.RecomputeFixedXFromCenter(ScreenData.QuestLog.ItemStartX)
-ScreenData.QuestLog.ItemStartY = Hephaistos.RecomputeFixedXFromCenter(ScreenData.QuestLog.ItemStartY)
+ScreenData.QuestLog.ItemStartY = Hephaistos.RecomputeFixedYFromCenter(ScreenData.QuestLog.ItemStartY)