+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   Fix a bunch of overlays (notably shops).
+
 ## [v0.5.0] - 2021-08-15
 
 ### Added