+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 -   Fix a bunch of overlays (notably shops).
+-   Fix quest log (fates) screen.
 
 ## [v0.5.0] - 2021-08-15